Split tests and add failing tests for equality
