+++ tests/all_tests.rs
@@ -1,0 +1,5 @@
+pub mod compiler;
+#[cfg(test)]
+pub mod helpers;
+pub mod r7rs;
+pub mod vm;
