+++ tests/r7rs/mod.rs
@@ -1,0 +1,2 @@
+pub mod equality_tests;
+mod storage_model_tests;
