+++ tests/vm/mod.rs
@@ -1,0 +1,5 @@
+pub mod closure_tests;
+pub mod language_tests;
+pub mod procedure_tests;
+pub mod regression_tests;
+pub mod smoke_tests;
